requirements.txt
supabase
flask
python-dotenv
gunicorn
flask_cors
numpy
twilio